Pull SummaryData component from Summary
