+++ context/app/static/js/components/Detail/Summary/Summary.jsx
@@ -2,21 +2,11 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Typography from '@material-ui/core/Typography';
 
-import {
-  FlexContainer,
-  FlexCenterAlign,
-  FlexColumnRight,
-  FlexRight,
-  StyledPaper,
-  StyledFileIcon,
-  JsonButton,
-  StyledDiv,
-} from './style';
+import SummaryData from '../SummaryData';
+import { FlexColumnRight, StyledPaper, StyledDiv } from './style';
 import SectionHeader from '../SectionHeader';
 import SectionItem from '../SectionItem';
-import SummaryItem from '../SummaryItem';
 import SectionContainer from '../SectionContainer';
-import StatusIcon from '../StatusIcon';
 
 function Summary(props) {
   const {
@@ -41,20 +31,9 @@ function Summary(props) {
       </Typography>
       <StyledDiv>
         <SectionHeader isSummary>{display_doi}</SectionHeader>
-        <FlexCenterAlign>
-          {entity_type !== 'Donor' && <FlexContainer>{children}</FlexContainer>}
-          <FlexRight>
-            {entity_type === 'Dataset' && status && status.length > 0 && (
-              <FlexCenterAlign>
-                <StatusIcon status={status} />
-                <SummaryItem>{status}</SummaryItem>
-              </FlexCenterAlign>
-            )}
-            <JsonButton href={`/browse/${entity_type.toLowerCase()}/${uuid}.json`} target="_blank" component="a">
-              <StyledFileIcon color="primary" />
-            </JsonButton>
-          </FlexRight>
-        </FlexCenterAlign>
+        <SummaryData entity_type={entity_type} uuid={uuid} status={status}>
+          {children}
+        </SummaryData>
       </StyledDiv>
 
       <StyledPaper>
