+++ context/app/static/js/components/Detail/SummaryData/index.js
@@ -1,0 +1,3 @@
+import SummaryData from './SummaryData';
+
+export default SummaryData;
